+++ Source/System/State/Base/OSGTextureObjChunk.h
@@ -256,7 +256,7 @@ class OSG_SYSTEM_DLLMAPPING TextureObjChunk : public TextureObjChunkBase
            UInt32 handleGL       (DrawEnv                 *pEnv, 
                                   UInt32                   id, 
                                   Window::GLObjectStatusE  mode,
-                                  UInt32                   uiOptions);
+                                  UInt64                   uiOptions);
 
     static void   handleDestroyGL(DrawEnv                 *pEnv, 
                                   UInt32                   id,
